# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  fy.RqCreateFY:
    properties:
      fy_year:
        type: string
      is_active:
        type: boolean
      label:
        type: string
    required:
    - fy_year
    - label
    type: object
  fy.RqUpdateFYLabel:
    properties:
      is_active:
        type: boolean
      label:
        type: string
    type: object
  fy.RsFinancialYear:
    properties:
      end_date:
        type: string
      id:
        type: string
      label:
        type: string
      start_date:
        type: string
    type: object
  practitioner.RsPractitioner:
    properties:
      abn:
        type: string
      id:
        type: string
      user:
        $ref: '#/definitions/practitioner.RsUserInfo'
      verified:
        type: boolean
    type: object
  practitioner.RsUserInfo:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
...
  response.RsError:
    properties:
      code:
        type: integer
      detail:
        type: string
      message:
        type: string
    type: object
  setting.RqCreatePractitioner:
    properties:
      abn:
        maxLength: 20
        type: string
      user_id:
        type: string
      verified:
        type: boolean
    required:
    - user_id
    type: object
  setting.RqUpdatePractitioner:
    properties:
      abn:
        maxLength: 20
        type: string
      verified:
        type: boolean
    type: object
  setting.RqUpsertPractitionerSetting:
    properties:
      color:
        type: string
      logo:
        maxLength: 255
        type: string
      timezone:
        maxLength: 255
        type: string
    type: object
  setting.RsPractitioner:
    properties:
      abn:
        type: string
      created_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      verified:
        type: boolean
    type: object
  setting.RsPractitionerSetting:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: integer
      logo:
        type: string
      practitioner_id:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
    type: object
...
  util.RsList:
    properties:
      items: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Backend API for acareca
  title: Backend API
  version: 1.0.0
paths:
...
  /admin/create-fy:
    post:
      consumes:
      - application/json
      parameters:
      - description: Financial Year Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/fy.RqCreateFY'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/fy.RsFinancialYear'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Create a new financial year
      tags:
      - fy
  /admin/get-fys:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.RsList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Get all financial years
      tags:
      - fy
  /admin/get-quarters/{financial_year_id}:
    get:
      parameters:
      - description: Financial Year UUID
        in: path
        name: financial_year_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.RsList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Get all quarters for a specific financial year
      tags:
      - fy
...
  /admin/update-fy/{financial_year_id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Financial Year UUID
        in: path
        name: financial_year_id
        required: true
        type: string
      - description: Updated Label Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/fy.RqUpdateFYLabel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fy.RsFinancialYear'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Update the label of a financial year
      tags:
      - fy
...
  /practitioner:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.RsList'
      security:
      - BearerToken: []
      summary: List all practitioners
      tags:
      - practitioner
  /practitioner/{id}:
    get:
      parameters:
      - description: Practitioner ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/practitioner.RsPractitioner'
      security:
      - BearerToken: []
      summary: Get practitioner by ID
      tags:
      - practitioner
...
  /setting:
    delete:
      consumes:
      - application/json
      description: delete a practitioner
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Delete a practitioner
      tags:
      - setting
    get:
      consumes:
      - application/json
      description: get a practitioner by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/setting.RsPractitioner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Get a practitioner by ID
      tags:
      - setting
    patch:
      consumes:
      - application/json
      description: update a practitioner
      parameters:
      - description: Updated Practitioner Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/setting.RqUpdatePractitioner'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/setting.RsPractitioner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Update a practitioner
      tags:
      - setting
    post:
      consumes:
      - application/json
      description: Register a new practitioner in the system
      parameters:
      - description: Practitioner Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/setting.RqCreatePractitioner'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/setting.RsPractitioner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Create a new practitioner
      tags:
      - setting
  /setting/by-user/{user_id}:
    get:
      consumes:
      - application/json
      description: get a practitioner by user ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/setting.RsPractitioner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Get a practitioner by user ID
      tags:
      - setting
  /setting/list:
    get:
      consumes:
      - application/json
      description: list practitioners
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.RsList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: List practitioners
      tags:
      - setting
  /setting/setting:
    get:
      consumes:
      - application/json
      description: get a setting by practitioner ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/setting.RsPractitionerSetting'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Get a setting by practitioner ID
      tags:
      - setting
    put:
      consumes:
      - application/json
      description: upsert a setting by practitioner ID
      parameters:
      - description: Setting Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/setting.RqUpsertPractitionerSetting'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/setting.RsPractitionerSetting'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Upsert a setting by practitioner ID
      tags:
      - setting
schemes:
- http
- https
securityDefinitions:
  BearerToken:
    description: Type "Bearer <your_token>" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"